+++ src/infra/database/knex/migrations/20220304195701_create_product.ts
@@ -6,7 +6,7 @@ export async function up(knex: Knex): Promise<void> {
         table.increments();
         table.uuid('uuid').notNullable().unique();
         table.string('name').notNullable();
-        table.float('price').notNullable();
+        table.integer('price').notNullable();
         table.integer('available').notNullable();
         table.timestamp('created_at').defaultTo(knex.fn.now())
         table.timestamp('updated_at').defaultTo(knex.fn.now())
